디렉토리 구조 변경
/chat 주소로 바로 접근할 경우 navigation bar 없이 채팅 페이지만 뜨게
했습니다. 이를 위해 로그인해야 접근 가능한 페이지들 중, chat 페이지만을
제외한 나머지는 (nav) route group에 집어넣었습니다. (nav)/(.)chat에서
intercepting routing을 합니다.

login/page.tsx, (nav)/page.tsx, chat/page.tsx에 똑같은 main 태그가
반복되어 나중에 리팩토링할 여지가 있어보입니다
