+++ front/app/(main)/@home/(nav)/(.)chat/page.tsx
@@ -1,7 +1,5 @@
 import ChatLayout from "@/components/Chat/ChatLayout";
 
 export default function ModalPage() {
-    return (
-        <ChatLayout />
-    );
+    return <ChatLayout />;
 }
