+++ front/app/(main)/@home/(nav)/game/[[...id]]/page.tsx
