+++ front/app/(main)/@home/(nav)/social/page.tsx
