+++ front/app/(main)/@home/(nav)/welcome/page.tsx
