+++ front/app/(main)/@home/(nav)/welcome2/page.tsx
